--- a/app/models/vmdb_table.rb
+++ b/app/models/vmdb_table.rb
@@ -1,14 +1,12 @@
 class VmdbTable < ApplicationRecord
   belongs_to :vmdb_database
 
-  has_many :vmdb_indexes,                            :dependent => :destroy
+  has_many :vmdb_indexes                             # Destroy will be handled by seeder
   has_many :vmdb_metrics,          :as => :resource  # Destroy will be handled by purger
 
   has_one  :latest_hourly_metric,  -> { VmdbMetric.where(:capture_interval_name => 'hourly', :resource_type => 'VmdbTable', :timestamp => VmdbMetric.maximum(:timestamp)) }, :as => :resource, :class_name => 'VmdbMetric'
 
   include VmdbDatabaseMetricsMixin
-
-  include_concern 'Seeding'
 
   serialize :prior_raw_metrics