… app/models/vmdb_index/seeding.rb
… app/models/vmdb_table/seeding.rb
… app/models/vmdb_table_evm/seeding.rb
… app/models/vmdb_table_text/seeding.rb
… spec/models/vmdb_database_spec.rb
… spec/models/vmdb_index_spec.rb
… spec/models/vmdb_table_evm_spec.rb
… spec/models/vmdb_table_spec.rb