Initial seed dropped from 6.14s to 2.48s (60% savings)
Second seed dropped from 1.64s to 0.025s (98.5% savings)

Basically, the big change is that instead of ((N+1)+1)+1 querying Postgres for each table -> index -> (text_table -> index), we just make 2 queries - 1 for all tables including text_tables, and 1 for all indexes.  Then we just do a simple desired state loop over each.

@jrafanie Please review
cc @NickLaMuro