@Fryguy 2019-01-04
I kept all of the logging of each record, but I really don't like it.  I would love to remove the logging or make it less verbose, but curious what you think.