@NickLaMuro 2019-01-04
@Fryguy I will look at this later, after I get ~~lunch~~ dinner... but is this related, or a different approach to https://github.com/ManageIQ/manageiq/pull/18271  ?

cc @kbrock