@Fryguy 2019-01-04
@NickLaMuro Yes, looks similar, and I hadn't seen it.  Honestly, though, I like my approach better because it...

a) gets rid of the myriad seeding.rb files everywhere, and just does it in one place
b) doesn't create virtual attributes (which in turn means they aren't exposed to reporting to anything else); these things are *only* used at seed time, so not exposing them is preferable.
c) feels simpler to wrap my head around