@Fryguy 2019-01-05
> leave something to be desired in regards to documentation

Can you elaborate?  I feel like this code is pretty self-documenting, so I'm curious what more you would like to see?  I commented the new ar_dba methods at least so the expected return types are clear.