@Fryguy 2019-01-07
@NickLaMuro @jrafanie I implemented some of your suggestions.  Also found another perf improvement, so the second run dropped even more, was 0.56s (66%) and is now 0.25s (85%)