@Fryguy 2019-01-07
I goofed up the math on the OP.  I wrote that it went to 0.25s (85%), but it actually went down to 0.***0***25s (98.5%)!  Updated the commit message and OP accordingly

![mundane_detail](https://user-images.githubusercontent.com/52120/50790057-db450b00-128b-11e9-8135-94b48120e457.gif)